ValkyriaChroniclesFix: Initial commit with working fix
This Fix has two parts: the DLL and the python patcher. Given the nature
of this game and the huge amount of initialization done, the python
script patches the exe where the DLL could never be fast enough to
inject some code.
The DLL portion is pretty standard hooks into locations to change some
things around. Makes the game work in ultrawide reses and centers or
spans the HUD to your liking. The spanning of the HUD applies to
everything, thats just the way the game was made, its too bothersome
to attempt to isolate everything and just not worth it.

	new file:   .gitignore
	new file:   .gitmodules
	new file:   CMakeLists.txt
	new file:   LICENSE
	new file:   README.md
	new file:   ValkyriaChroniclesPatch.py
	new file:   inc/utils.hpp
	new file:   install.ps1
	new file:   safetyhook
	new file:   spdlog
	new file:   src/dllmain.cpp
	new file:   src/utils.cpp
	new file:   yaml-cpp
	new file:   zydis