ValkyriaChroniclesFix: Remove all things fov fix and bump version
The FOV is too engrained in various components of the game and
utimately not a worthwhile pursuit, the master FOV in the game not only
controls the FOV of the game model in game, but also the background
when viewing cutscenes since its the same camera just panned to a
completely different angle.

Bumped version to 1.0.1 to signify FOV removal.

	modified:   install.ps1
	modified:   src/dllmain.cpp